Addressed pull request suggestions made by @freeman-lab.  Moved unit tests out into test_decorators.py
